Move Contact block and add new Catalog page type
